feat: added routing using react-router
